Added split_list.py and GenoPheno.sh, fully automated workflow, updated README
GenoPheno.sh allows you to run the
entire workflow in one command,
using only a MIM number input list
that can be greater than 20, an
API key obtained from OMIM, and
a name for your project.

split_list.py allows you to split a list
of MIMs that is greater than 20 into
separate lists for use in table.py.
This can be used manually by the
user and it is also what GenoPheno.sh
uses to split a large list before
processing.