Fixed table.py error output and crashing due to MIMs with clinical data in the old OMIM format and variation in where OMIM stores gene data.
I believe the program retrieves the genes
associated with the MIM most of the time.  I am
still taking into consideration whether or not
I will want to make significant changes to the
code that will allow for the retrieval of genes
that are stored in the "geneMap" OMIM API
hierarchy, which is a different branch from the
"phenotypeMap" hierarchy, which is where the gene
data is sometimes located.  From what I've
observed, each MIM appears to store genetic data
in either one or the other in such a way that
geneMap and phenotypeMap branches are present
mutually exclusively of one another.

I will continue to test the program on various
batches of MIMs before considering dealing with
the fact that some MIMs use geneMap rather than
phenotypeMap since phenotypeMap appears to be
the main method of gene data storage.  I might
also e-mail OMIM to ask about this.