Adjusted graph.py to approximate (instead of calculate) average node connectivity.  This was done to massively speed up execution time in order to run several large experiments.  It is a temporary change, but in the future I think I will add an execution argument to allow for either approximating or calculating this metric.
Changed control_trials_generator.py to write .pbs
scripts that copy the original GitHub repository
instead of cloning it.  I'm not sure what I was
thinking before -- I believe this was a careless
logical mistake.

Also of note: control_trials_generator.py is not
quite ready for use by the public, even though
it has been published, because I still need to
add an argument at the beginning to specify
a working directory.