Version Update across suite: 7.1 -> 7.2
Fixed interactors.py crash that
was due to issues regarding how
"failedIds" was returned from the
API call for gene MIM ID conversion
to UNIPROT ID.  If there were no
failed IDs, the "failedIds" list
was not returned, whereas before
I had thought that it was always
returned with an empty list if
there were no failed IDs.

Also, I completely fixed the way
the mapped lists coordinate with
more solid code and completely
changed the output to reflect this.

Output now features UNIPROT ID for
each gene, and ENSEMBL ID has been
eliminated.  This was done to better
convey what the program is doing
and to do this as transparently
as possible.