Combined the two OMIM API calls in table.py into a single API call and edited some code to reconcile these changes with the way the program was originally designed to parse the OMIM data.
Output appears to be the same as
in the last committed version.

If testing on the Palmetto Cluster
validates that there are no issues,
I will do another commit to update
the version number accross the
repository.

I will have to clean up some of the
commenting later if a reversion
is not necessary.